feat:copies code from files for fix errors, and remove this files
