--- assets/focusArea.js
@@ -1,8 +1,0 @@
-
-
-function focusArea(){
-    const textarea = document.querySelector(".display");
-    textarea.focus();
-}
-
-export default focusArea;
