--- assets/langEn.js
@@ -1,15 +1,0 @@
-export default {
-	lang: 'en',
-	normal: [
-	  'Esc 1 2 3 4 5 6 7 8 9 0 - = bksp tab q w e r t y u i o p [ ] \\ caps a s d f g h j k l ; , enter shift z x c v b n m , . / ᐃ shift ctrl alt ` space en alt ᐊ ᐁ ᐅ del',
-	],
-	shift: [
-	  'Esc ! @ # $ % ^ & * ( ) _ + bksp tab Q W E R T Y U I O P { } | caps A S D F G H J K L : " enter shift Z X C V B N M < > ? ᐃ shift ctrl alt ` space en alt ᐊ ᐁ ᐅ del',
-	],
-	caps: [
-	  'Esc 1 2 3 4 5 6 7 8 9 0 - = bksp tab Q W E R T Y U I O P [ ] \\ caps A S D F G H J K L ; , enter shift Z X C V B N M , . / ᐃ shift ctrl alt ` space en alt ᐊ ᐁ ᐅ del',
-	],
-	capsOnShift: [
-	  'Esc ! @ # $ % ^ & * ( ) _ + bksp tab q w e r t y u i o p { } | caps a s d f g h j k l : " enter shift z x c v b n m < > ? ᐃ shift ctrl alt ` space en alt ᐊ ᐁ ᐅ del',
-	],
- }
