--- assets/langRu.js
@@ -1,15 +1,0 @@
-export default {
-	lang: 'ru',
-	normal: [
-	  'Esc 1 2 3 4 5 6 7 8 9 0 - = bksp tab й ц у к е н г ш щ з х ъ \\ caps ф ы в а п р о л д ж э enter shift я ч с м и т ь б ю . ᐃ shift ctrl alt ё space ru alt ᐊ ᐁ ᐅ del',
-	],
-	shift: [
-	  'Esc ! " № ; % : ? * ( ) _ + bksp tab Й Ц У К Е Н Г Ш Щ З Х Ъ / caps Ф Ы В А П Р О Л Д Ж Э enter shift Я Ч С М И Т Ь Б Ю , ᐃ shift ctrl alt Ё space ru alt ᐊ ᐁ ᐅ del',
-	],
-	caps: [
-	  'Esc 1 2 3 4 5 6 7 8 9 0 - = bksp tab Й Ц У К Е Н Г Ш Щ З Х Ъ \\ caps Ф Ы В А П Р О Л Д Ж Э enter shift Я Ч С М И Т Ь Б Ю . ᐃ shift ctrl alt Ё space ru alt ᐊ ᐁ ᐅ del',
-	],
-	capsOnShift: [
-	  'Esc ! " № ; % : ? * ( ) _ + bksp tab й ц у к е н г ш щ з х ъ / caps ф ы в а п р о л д ж э enter shift я ч с м и т ь б ю , ᐃ shift ctrl alt ё space ru alt ᐊ ᐁ ᐅ del',
-	],
- }
