--- assets/mouseClickDown.js
@@ -1,7 +1,0 @@
-function mouseClickDown(event) {
-	const key = event.target.closest(".key");
-	if (!key) return;
-	key.classList.add("active");
-}
-
-export default mouseClickDown;
