--- assets/mouseClickUp.js
@@ -1,7 +1,0 @@
-function mouseClickUp(event) {
-	const key = event.target.closest(".key");
-	if (!key) return;
-	key.classList.remove("active");
-}
-
-export default mouseClickUp;
